+++ source/TestCommon/source/FunctionApp.TestCommon/Configuration/AzureB2CSettings.cs
@@ -34,7 +34,7 @@ namespace Energinet.DataHub.Core.FunctionApp.TestCommon.Configuration
         public string BackendServicePrincipalObjectId { get; internal set; }
             = string.Empty;
 
-        public string BackendObjectId { get; internal set; }
+        public string BackendAppObjectId { get; internal set; }
             = string.Empty;
     }
 }