+++ source/TestCommon/source/FunctionApp.TestCommon/Configuration/IntegrationTestConfiguration.cs
@@ -104,7 +104,7 @@ namespace Energinet.DataHub.Core.FunctionApp.TestCommon.Configuration
                 ServicePrincipalSecret = configuration.GetValue("AZURE-B2C-SPN-SECRET"),
                 BackendAppId = configuration.GetValue("AZURE-B2C-BACKEND-APP-ID"),
                 BackendServicePrincipalObjectId = configuration.GetValue("AZURE-B2C-BACKEND-SPN-OBJECTID"),
-                BackendObjectId = configuration.GetValue("AZURE-B2C-BACKEND-OBJECTID"),
+                BackendAppObjectId = configuration.GetValue("AZURE-B2C-BACKEND-APP-OBJECTID"),
             };
         }
     }